Added BaseTest class with @before and @after. MakeOrder and CheckQAAParametrized are updated to extend BaseTest to avoid code duplication in tests.
