+++ src/test/java/ru/yandex/prakticum/main/BaseTest.java
@@ -1,0 +1,31 @@
+package ru.yandex.prakticum.main;
+
+import org.junit.After;
+import org.junit.Before;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.firefox.FirefoxDriver;
+
+import java.time.Duration;
+
+public class BaseTest {
+    protected WebDriver driver;
+
+    @Before
+    public void setUp() {
+        System.setProperty("webdriver.chrome.driver", "C:\\WebDriver\\bin\\chromedriver.exe");
+        System.setProperty("webdriver.gecko.driver", "C:\\WebDriver\\bin\\geckodriver.exe");
+
+        // выбор браузера для прогона тестов:
+        driver = new ChromeDriver();
+        //driver = new FirefoxDriver();
+
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));
+    }
+
+    @After
+    public void tearDown() {
+        driver.quit();
+    }
+}
+
